Build basic IPC structure for the actual music playback handling
Since audio is finicky this is probably for the best generally, but
we're specifically doing this to workaround the `speaker` library
printing info to stderr that we can't seem to suppress; by putting this
stuff in a separate process, we have full control over stdio and can
also (hopefully) gracefully recover if something terrible happens.

It's definitely a bummer, but probably will still use fewer resources
than a full electron app, so... it's okay?