Add preliminary native macOS component to support media keys, etc. (#2)
* Add proof of concept native macOS app

Hopefully we'll be able to use this via IPC to provide native media key
integrations for macOS, as well as show media info

* Set up basic IPC event flow from native -> cljs

* Forward :toggle native event to re-frame as :player/play-pause

* Mark native app as "agent" so it doesn't create a dock icon

* Add app -> native RPC messaging

* Wire up events to native IPC

* Fix some compile warnings with type hints

* Improve robustness of the native component + fix interactions w/it

In particular, even with flushing between each message we might still
receive multiple messages in one "data" event chunk, so we have to split
those up to receive them properly. In general this shouldn't be an
issue, but helpful in dev in the presence of logging, for example

* Don't create the system tray icon, for now

The code remains in case we find a use for it in the future, but it's
nice that we don't *need* it, since it doesn't do anything at the
moment.