Implement seekable AudioTrack abstraction
AudioTrack is sort of a state machine, like similar abstractions
provided by platforms, and built on lower-level machinery. In
particular, it utilizes an IPCMSource to be able to repeatedly reopen
a stream from *somewhere* that it can then skip through as necessary
(based on the configuration of the PCM stream) to jump to the right
spot in a track, based on a specified `seek`.

Under the hood, it hands this PCM stream to an AudioClip, which handles
routing the audio to a speaker, adjusting volume, etc. Since an
AudioClip is one stream to one speaker, it cannot seek nor switch output
device. These functions are meant to be handled by the AudioTrack by
simply closing the old AudioClip and creating a new one.