Rewrite playback handling with seekable, managed AudioTrack (#14)
* Implement seekable AudioTrack abstraction

AudioTrack is sort of a state machine, like similar abstractions
provided by platforms, and built on lower-level machinery. In
particular, it utilizes an IPCMSource to be able to repeatedly reopen
a stream from *somewhere* that it can then skip through as necessary
(based on the configuration of the PCM stream) to jump to the right
spot in a track, based on a specified `seek`.

Under the hood, it hands this PCM stream to an AudioClip, which handles
routing the audio to a speaker, adjusting volume, etc. Since an
AudioClip is one stream to one speaker, it cannot seek nor switch output
device. These functions are meant to be handled by the AudioTrack by
simply closing the old AudioClip and creating a new one.

* Follow the default speaker on AudioTrack.play

* Reorganize, and scaffold out support for CachingPCMSource

* Add preliminary caching implementations

Re-opening the stream while the download is in progress doesn't seem to
be working very well...

* Finally (!) fix the new multi-openable CachingPCMSource

* Add EventfulAudioTrack to implement IPlayable; use it!

* Remove old, unused player code

* Use a transform instead of (.on "data") to count bytes read

Setting (.on "data") immediately transforms the stream into flowing
mode, so we might actually lose some bytes doing that.

* Refactor to use more reusable, composable Transforms

* Fix incorrect double-piping

* Improve partial-pipe robustness and reduce CPU churn

We no longer spam the disk trying to constantly trying to re-open the
file until it happens to have more data, and instead react to updates to
the state atom so we *know* there is more data before trying.

* Switch to reading from the moved file when the download completes

* Move analyze-caching to a public fn in the analyze ns

Also adds some docs to pcm.caching

* Utilize parallel-let (p/plet) in a few places where it makes sense