Ensure that (IAudioTrack/play) is a nop after (close)
See #32

Haven't ben able to reproduce the bug but this seems like the most
likely candidate as the cause. My suspicion is that one of the async
processes kicked off in (play) is taking some time, and in the meantime
(close) gets called when the track is swapped. Since there were no
checks of the :closed? state when they return, it creates the clip
anyway and starts playing.