Handle "system" errors on play by restarting with a new clip
Fixes #31

This is a bit tricky because the error is provided via a "threadsafe"
callback which puts the callback-throw on a queue, and we don't have a
similar callback-provided positive success assertion, so we have to use
a timeout to just wait for some time and *hope* that the callback wasn't
very slow in firing.

This in turn means that IAudioClip/play now returns a Promise instead of
being simple and synchronous. A bummer to be sure, but not the end of
the world.