ytm: Fix home page loading
It's unclear why the YTMusic parsing broke, but we now basically just
use the same processing as we already have for many other things with
very few changes, which is nice---less code to maintain.

This also enables support for single songs within other
carousels/shelves, since we needed it here.